Tons of stuff
- Implement reading and editing in file manager
- Use zerolog in file and project managers
- Update log format

+++ pkg/languageserver/process.go
@@ -55,6 +55,9 @@ func (p *ProcessImpl) Start() error {
 }
 
 func (p *ProcessImpl) Stop() error {
+	if p.cmd.Process == nil {
+		return nil
+	}
 	return p.cmd.Process.Kill()
 }
